Add mirrored test-suite tests
